+++ packages/core/src/components/accordion/components/handles/index.ts
@@ -1,4 +1,0 @@
-export * from "./element.handle";
-export * from "./heading.handle";
-export * from "./icon.handle";
-export * from "./state-icon.handle";