+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ### Added
 * Graphics: Foutmelding ([#2388](https://github.com/dso-toolkit/dso-toolkit/issues/2388))
+* Accordion: `onDsoAnimationStartEvent`, instant `scrollIntoView()`, `hasNestedAccordion` prop en `AccordionSection.focusHandle()` ([#2408](https://github.com/dso-toolkit/dso-toolkit/issues/2408))
 
 ### Fixed
 * Date Picker: Input en output worden niet prepadded met "0" ([#2409](https://github.com/dso-toolkit/dso-toolkit/issues/2409))